# optimize data structures for 5.5x speedup