This commit improves the layout of the "parser transition state" and
"grapheme code point" data structures; changes them to use structs
instead of arrays; and adds an ASCII and Extended-ASCII fast path for
the property() function.

Overall, this results in a ~5.5x speedup for Graphemes.Next().

```
goos: darwin
goarch: arm64
pkg: github.com/rivo/uniseg

name                        old time/op    new time/op    delta
GraphemesClass-10             2.87µs ± 0%    0.76µs ± 0%  -73.50%  (p=0.000 n=10+8)
GraphemesNext-10              2.55µs ± 0%    0.47µs ± 0%  -81.75%  (p=0.000 n=8+9)
GraphemesFunctionBytes-10     0.64ns ± 0%    0.64ns ± 0%     ~     (p=0.344 n=9+10)
GraphemesFunctionString-10    0.64ns ± 0%    0.64ns ± 0%     ~     (p=0.190 n=10+9)

name                        old alloc/op   new alloc/op   delta
GraphemesClass-10               944B ± 0%      944B ± 0%     ~     (all equal)
GraphemesNext-10               0.00B          0.00B          ~     (all equal)
GraphemesFunctionBytes-10      0.00B          0.00B          ~     (all equal)
GraphemesFunctionString-10     0.00B          0.00B          ~     (all equal)

name                        old allocs/op  new allocs/op  delta
GraphemesClass-10               3.00 ± 0%      3.00 ± 0%     ~     (all equal)
GraphemesNext-10                0.00           0.00          ~     (all equal)
GraphemesFunctionBytes-10       0.00           0.00          ~     (all equal)
GraphemesFunctionString-10      0.00           0.00          ~     (all equal)
```

*Note*: The GraphemesNext benchmark was backported to d6773baf for these
comparisons.