@rivo 2022-07-22
I'll write this here as there is no issue to go along with the PR's you submitted and I'd like to note a few general things:

- First of all, thank you for contributing to this project!
- I'm currently in the process of making major changes to `uniseg`. (You can see some of the process in the `wordbreak` branch and there will be other branches, too.)
- These changes are so extensive, you could almost call them a rewrite of the package.
- Therefore, by now, many of the PR's you submitted are clashing with those changes. It's not your fault, I didn't tell anyone that I'm working on this right now.
- I should say I prefer if people open issues first, before submitting PR's. I outline this [here](https://github.com/rivo/tview/blob/master/CONTRIBUTING.md). Again, not your fault, because that guide is not in `uniseg` (mostly because it doesn't get a lot of contributions so there's less guidance from my side).

In any case, once I've finish my current work on `uniseg`, you're welcome to change your PR's accordingly — although again, I would prefer that we discuss the proposed changes before you spend time writing code that may not end up being used.