@rivo 2022-07-23
I should also note that if you want to keep working on this PR after the changes mentioned above, it should be broken into multiple PR's so we can discuss/approve each of these changes separately. For example, there are changes in the generators (not really relevant), the rule map has been reordered, sacrificing readability (very likely to be rejected), some suggestions are not helpful (e.g. turning `i / 2` into `int((uint) i >> 1)`), and others are certainly useful. But with all of them in one PR, I can only merge, close, or we go back and forth on each individual change with review comments. I probably don't have time for the latter.